requirements.txt
# Core dependencies for Pneumonia Detection System
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
python-multipart>=0.0.6

# PyTorch and vision libraries
torch>=2.0.0
torchvision>=0.15.0

# Machine learning utilities
scikit-learn>=1.3.0
numpy>=1.24.0

# Type hints
typing-extensions>=4.8.0
